Re-format the integration test files
This formats all the integration test source files with IntelliJ's
default Groovy configuration options.

If there's some purpose-made formatter like clang-format for Groovy
that'd be ideal, but as of mid 2021 I'm not aware of one and quite a bit
of the formatting looks like it was written without using an
auto-formatter.